\name{Plotting functions}
\alias{image2D.TempSEDdyn}
\alias{matplot1D}
\alias{matplot1D.default}
\alias{matplot1D.TempSEDdyn}
\alias{matplot0D}
\alias{matplot0D.default}
\alias{matplot0D.TempSEDdyn}
\alias{plot.TempSEDdyn}

\title{
  Utility functions for plotting sediment temperature dynamic model output.
}
\description{

  \code{image2D} generates a 2-D image (x=time, y=depth) of the dynamic output (temperature (color variable) versus time (x) and depth (y)).

  \code{matplot1D} plots vertical temperature profiles versus depth.

  \code{plot} plots the output of 0D variables (one figure each).
  
}


\usage{

\method{image2D}{TempSEDdyn}(z, which = NULL, subset = NULL,
    time_unit = c("sec", "hour", "day", "year"), 
    xlab = time_unit, time_origin = NULL, plotWater = FALSE,
   colkey = list(cex.clab = 0.8, line.clab = 0.5, cex.axis = 0.8), 
   \dots)

matplot0D (x, ...)
\method{matplot0D}{default}(x, ...)
\method{matplot0D}{TempSEDdyn}(x, ..., select = NULL, which = select, 
                            time_unit = c("sec", "hour", "day", "year"), 
                            xlab = time_unit, lty = 1)

matplot1D (z, ...)
\method{matplot1D}{default}(z, ...)
\method{matplot1D}{TempSEDdyn}(z, 
    type = "l", col = "grey", lty = 1, \dots)

\method{plot}{TempSEDdyn}(x, \dots, select = NULL, which = select, 
     time_unit = c("sec", "hour", "day", "year"), 
     xlab = time_unit, ylab = NULL, lty = 1, las = 1)
}
\arguments{

  \item{x, z}{ object of class \code{TempSED} generated by \link{TempSED_run1D}.
  }
  \item{select, which}{ The name(s) of the 1-dimensional output to be plotted versus time.
  }
  \item{subset}{ either a logical expression indicating elements or rows to keep, or a vector of integers denoting the indices of the rows to keep. Missing values are taken as FALSE
  }
  \item{time_origin}{ When not \code{NULL}, will use time labels rather than seconds to label time. This may not work well with \code{time_unit}; (use Sys.setlocale(LC_TIME) to change how the date appears on the x-axis).
  }
  \item{time_unit}{ The time unit of the x-axis. The default is "sec" (the unit of the model).
  }
  \item{plotWater}{ When TRUE, will also plot the properties of the overlying watercolumn (height and temperature).
  }
  \item{type, col, lty, xlab, ylab, las}{ parameters passed to the plot function
  }
  \item{colkey}{ A list with specifications of the color key. The default is to have a smaller  title (\code{cex.clab}) and axis labels (\code{cex.axis}), the key title positioned close to the color key (\code{line.clab}), See \link[plot3D]{colkey}
  }
  \item{...}{ Any argument passed to the functions \link[plot3D]{image2D} .
  }
}


\author{
  Karline Soetaert
}


\examples{

fAirTempYr    <- function(t) 
                    return(15 + 15*sin(2*pi*t/86400/365))      # dgC
fSolarRadYr   <- function(t) 
                    return(150*(1+0.8*sin(2*pi*t/86400/365)))  # W/m2
times         <- seq(from = 0, to = 365*86400, by = 86400)

Tout  <- TempSED_run1D(
                   T_ini            = 10, 
                   times            = times, 
                   f_Airtemperature = fAirTempYr,
                   f_Solarradiation = fSolarRadYr)
image2D(Tout)

}

\seealso{
  \code{\link{TempSED_run1D}}.
  \code{\link{TempSED_get0Dvars}}.
  \code{\link{TempSED_getparms}}.
}

\keyword{ utilities }
